apps/astro-engine/pyproject.toml
[project]
name = "astro-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pyswisseph",
    "pydantic",
    "python-dateutil",
    "google-genai>=1.74.0",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
